// type.ts
export namespace TMDBResult {
  export interface Result {
    adult: boolean;
    backdrop_path: string;
    genre_ids: number[];
    id: number;
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string;
    release_date: string;
    title: string;
    video: boolean;
    vote_average: number;
    vote_count: number;
  }

  export interface RootObject {
    page: number;
    results: Result[];
    total_pages: number;
    total_results: number;
  }
}

export namespace TMDBGenreResult {
  export interface Genre {
    id: number;
    name: string;
  }

  export interface RootObject {
    genres: Genre[];
  }
}